webserver.catalog: renamed modules to reduce coupling
